Optimizations (#47)
* Added a **Use Case Roadmap** section to `README.md` to provide better project direction and usage visibility.
* Refined `PInvoke.BootstrapBufferStorage` feature switch handling for improved configurability and maintainability.
* Set `StandardStorage` as the default storage implementation across all supported target frameworks.
* Added `NativeUtilities.HeapAlloc<T>(Int32)` to simplify native memory allocation scenarios.
* Optimized several extension methods to improve performance and reduce overhead.

+++ src/Intermediate/Rxmxnx.PInvoke.Buffers.Intermediate/BufferManager.cs
@@ -35,7 +35,7 @@ public static partial class BufferManager
 #if !NET8_0_OR_GREATER
 		=> UInt16.MaxValue;
 #else
-		=> MetadataStorage.MaxCapacity == 0 ? UInt16.MaxValue : MetadataStorage.MaxCapacity;
+		=> MetadataStorage.MaxCapacity == 0 ? UInt16.MaxValue : (UInt16)MetadataStorage.MaxCapacity;
 #endif
 
 	/// <summary>
